Streamlit/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
tensorflow.keras
os
PIL

